Move SeparableConv2DQuantized to helper file for better discoverability.
PiperOrigin-RevId: 534168938
